kice-pass-gen/requirements.txt
PyQt6
fpdf2
requests
